Added schema extraction methods, and utilities for transforming NWB files
The schema extraction methods are for things like finding all the numeric attributes of nodes or edges.  This is useful in mutateParameters().

The system for transforming NWB files is good for doing things like adding or deleting a node attribute, calculating an attribute based on the other ones, and inserting or deleting nodes.  It's contained in the edu.iu.nwb.util.nwbfile.pipe package, and the main starting point is ParserPipe.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5574 c287de05-5010-0410-818a-8f02fb49d6a9