Updated Ghostscript-based Postscript conversion plugin.  We now provide converter algorithms for (1) saving PS as PDF, (2) handling PDF, (3) saving PS as PNG, and (4) handling PNG.
This plugin requires an external, PATH-visible Ghostscript installation.  I have tested this using GPL Ghostscript 9.02 for 32-bit Windows.  The plugin is designed to detect a Ghostscript interpreter on the PATH named "gswin64c.exe" or "gswin32c.exe" (on Windows systems) or "gs" (on Unix-based systems).

The "util" subpackage contains new FileFormat types for general work with Data and Files.  These could be moved to a more central location later.

Reviewed by Thomas.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5634 c287de05-5010-0410-818a-8f02fb49d6a9